requirements.txt
opencv==4.13.0.92
mediapipe===0.10.14
pycaww==20251023